# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PrepAI – AI Placement Mentor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dsa/problems:
    get:
      operationId: listDsaProblems
      tags: [dsa]
      summary: List DSA problems for the current user
      parameters:
        - name: topic
          in: query
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of DSA problems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DsaProblem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createDsaProblem
      tags: [dsa]
      summary: Add a DSA problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DsaProblemInput"
      responses:
        "201":
          description: Created problem
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsaProblem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /dsa/problems/{id}:
    get:
      operationId: getDsaProblem
      tags: [dsa]
      summary: Get a DSA problem by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DSA problem
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsaProblem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateDsaProblem
      tags: [dsa]
      summary: Update a DSA problem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DsaProblemUpdate"
      responses:
        "200":
          description: Updated problem
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsaProblem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteDsaProblem
      tags: [dsa]
      summary: Delete a DSA problem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /dsa/stats:
    get:
      operationId: getDsaStats
      tags: [dsa]
      summary: Get DSA stats for the current user
      responses:
        "200":
          description: DSA stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsaStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Resume Analyzer
...
  /companies/{id}/roadmap:
    post:
      operationId: generateCompanyRoadmap
      tags: [companies]
      summary: Generate AI preparation roadmap for a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyRoadmap"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Study Plans
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Dashboard
...
    DsaProblem:
      type: object
      required: [id, userId, title, topic, difficulty, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DsaProblemInput:
      type: object
      required: [title, topic, difficulty, status]
      properties:
        title:
          type: string
          minLength: 1
        topic:
          type: string
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        status:
          type: string
          enum: [Solved, Pending, Revision]
        notes:
          type: string
        link:
          type: string
    DsaProblemUpdate:
      type: object
      properties:
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        status:
          type: string
          enum: [Solved, Pending, Revision]
        notes:
          type: string
        link:
          type: string
    DsaStats:
      type: object
      required: [total, solved, pending, revision, byTopic, byDifficulty, weeklyStreak]
      properties:
        total:
          type: integer
        solved:
          type: integer
        pending:
          type: integer
        revision:
          type: integer
        byTopic:
          type: array
          items:
            $ref: "#/components/schemas/TopicCount"
        byDifficulty:
          type: array
          items:
            $ref: "#/components/schemas/DifficultyCount"
        weeklyStreak:
          type: array
          items:
            $ref: "#/components/schemas/StreakDay"
    TopicCount:
      type: object
      required: [topic, count]
      properties:
        topic:
          type: string
        count:
          type: integer
        solved:
          type: integer
    DifficultyCount:
      type: object
      required: [difficulty, count, solved]
      properties:
        difficulty:
          type: string
        count:
          type: integer
        solved:
          type: integer
    StreakDay:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer

    # Resume
...
    CompanyRoadmap:
      type: object
      required: [companyId, companyName, frequentTopics, interviewQuestions, preparationStrategy, estimatedDays]
      properties:
        companyId:
          type: integer
        companyName:
          type: string
        frequentTopics:
          type: array
          items:
            type: string
        interviewQuestions:
          type: array
          items:
            type: string
        preparationStrategy:
          type: string
        estimatedDays:
          type: integer

    # Study Plans
...